+++ src/app/movies/movies.component.ts
@@ -1,10 +1,14 @@
 import { Component, ElementRef, HostListener, Renderer2 } from '@angular/core';
-import { FavouriteType, MovieType, UserFavouriteType } from '../shared/types/movie.type';
+
+// service
 import { GlobalService } from '../shared/services/global.service';
-import { GlobalMovieService } from '../shared/services/global-movie/movie.service';
+import { GlobalMovieService } from '../shared/services/global-movie/global-movie.service';
 import { SnackbarService } from '../shared/template/snackbar/snackbar.service';
-import { UserType } from '../shared/types/auth.type';
 import { MoviesService } from './movies.service';
+
+// type
+import { FavouriteType, MovieType, UserFavouriteType } from '../shared/types/movie.type';
+import { UserType } from '../shared/types/auth.type';
 
 @Component({
   selector: 'app-movies',
@@ -15,11 +19,16 @@ export class MoviesComponent {
   showNavbar = true;
   elementsArray!: NodeListOf<Element>;
   userInfo!: UserType;
+  reachedBottom: boolean = false
 
-  TopRatedMovies!: MovieType[]
+  TopRatedMovies: MovieType[] = []
+  newTopRatedMovies!: MovieType[]
 
   favourite!: FavouriteType;
   usersData!: UserFavouriteType
+
+  currentPage = 1
+  loading = false
 
   constructor(
     private globalService: GlobalService,
@@ -34,17 +43,32 @@ export class MoviesComponent {
   }
 
   ngOnInit() {
-    this.getTopRatedMovies()
-    setTimeout(() => {
-      this.elementsArray =
-        this.element.nativeElement.querySelectorAll('.animated-fade-in');
-      this.fadeIn();
-    }, 500);
+    this.getTopRatedMovies(this.currentPage)
   }
 
   @HostListener('window:scroll', ['$event'])
   onScroll() {
     this.fadeIn();
+
+    const windowHeight = window.innerHeight + 20;
+    const scrollY = window.scrollY;
+    const bodyHeight = document.body.offsetHeight;
+
+    if (windowHeight + scrollY >= bodyHeight) {
+      this.fetchNextPage();
+    }
+  }
+
+  fetchNextPage() {
+    if (!this.loading) {
+      this.loading = true;
+
+      this.currentPage++;
+
+      this.getTopRatedMovies(this.currentPage).then(() => {
+        this.loading = false;
+      });
+    }
   }
 
   fadeIn() {
@@ -58,36 +82,50 @@ export class MoviesComponent {
     }
   }
 
-  async getTopRatedMovies() {
+  async getTopRatedMovies(pageNumber: number) {
     this.snackbar.showLoading(true)
     try {
-      const movies = await this.moviesService.topRatedMovies();
-      this.TopRatedMovies = movies.results;      
+      const movies = await this.moviesService.getTopRatedMovies(pageNumber);
+      this.newTopRatedMovies = movies.results;
+
+      if (this.newTopRatedMovies.length === 0) {
+        return;
+      }
+      this.TopRatedMovies.push(...this.newTopRatedMovies);
+      setTimeout(() => {
+        this.elementsArray =
+          this.element.nativeElement.querySelectorAll('.animated-fade-in');
+        this.fadeIn();
+      }, 500);
     } catch (error) {
       console.error('Error fetching movies:', error);
-      // this.snackbar.show('Error fetching movies');
+      this.snackbar.show('Error fetching movies');
       this.snackbar.showLoading(false)
     } finally {
-      console.log('API call completed.');
       this.snackbar.showLoading(false)
     }
   }
 
   async toggleFavorite(showId: number, mediaType: string): Promise<void> {
-    this.snackbar.showLoading(true)
-    try {
-      if (this.isFavorite(showId, mediaType)) {
-        const favourite = await this.movieService.removeFavourite(this.userInfo.uuid, showId, mediaType);
-        this.usersData = this.globalService.getUsersData();
-      } else {
-        const favourite = await this.movieService.addFavourite(this.userInfo.uuid, showId, mediaType);
-        this.usersData = this.globalService.getUsersData();
+    if (this.globalService.getToken()) {
+      this.snackbar.showLoading(true)
+      try {
+        if (this.isFavorite(showId, mediaType)) {
+          const favourite = await this.movieService.removeFavourite(this.userInfo.uuid, showId, mediaType);
+          this.usersData = this.globalService.getUsersData()
+        } else {
+          const favourite = await this.movieService.addFavourite(this.userInfo.uuid, showId, mediaType);
+          this.usersData = this.globalService.getUsersData()
+        }
+      } catch (error) {
+        console.error('Error toggling favorite:', error);
+        this.snackbar.show('Error toggling favorite');
       }
-    } catch (error) {
-      console.error('Error toggling favorite:', error);
-      // this.snackbar.show('Error toggling favorite');
-    } finally {
-      this.snackbar.showLoading(false)
+      finally {
+        this.snackbar.showLoading(false)
+      }
+    } else {
+      this.snackbar.show('Need to login first');
     }
   }
